+++ CHANGELOG.rst
@@ -9,6 +9,13 @@
 ERRERS: Change Log
 ==================
 
+Unreleased
+==========
+
+Fixed
+-----
+- Refer to project page at Python packaging index in help window.
+
 3.2.1, 2024-03-06
 =================